// src/app/shared/types/components.types.ts
export type PaleteVariant = 'LCA_PALETE' | 'GREEN_PALETE' | 'PURPLE_PALETE';

export type AlignVariant = 'left' | 'right' | 'center';

export type SizeVariant = 'extra-large' | 'large' | 'standard' | 'small' | 'xsmall' ;

export type TextVariant = 'uppercase' | 'capitalize' | 'lowercase' | 'none' ;

export type GenderVariant = 'f' | 'm';

export type PredictiveType = 'contains' | 'start' | 'none';

export type ButtonVariant =
  | 'primary'
  | 'secondary'
  | 'tertiary'
  | 'toggle'
  | 'action-primary'
  | 'action-secondary'
  | 'circle-primary'
  | 'circle-secondary'
  | 'icon-button';